Run OOC tests in one group
